fixup! perf(core): make sanitization tree-shakable in Ivy mode
