+++ packages/core/src/core.ts
@@ -34,5 +34,6 @@ export {EventEmitter} from './event_emitter';
 export {ErrorHandler} from './error_handler';
 export * from './core_private_export';
 export * from './core_render3_private_export';
-export {Sanitizer, SecurityContext} from './sanitization/security';
+export {SecurityContext} from './sanitization/security';
+export {Sanitizer} from './sanitization/sanitizer';
 export * from './codegen_private_exports';
